+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMPYE81SHJ5B1Y21QFDENWPN"
+        "01KMQK8T9ZAV01XYWMW631M796"
     )
     r"""The unique identifier of the data source"""