+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMPYE81VHCMQA2NWWBJ5JC87"
+        "01KMQK8TA1233PQKBJ5QNYVAH9"
     )
     r"""The unique identifier of the data source"""