+++ packages/orq-rc/RELEASES.md
@@ -6639,3 +6639,13 @@ Based on:
 - [python v4.7.0-rc.23] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.23] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.23 - packages/orq-rc
+
+## 2026-03-27 12:13:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.24] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.24 - packages/orq-rc